requirements.txt
fastapi
uvicorn
google-cloud-firestore
python-dotenv